add review
not done

+++ src/main/java/com/entities/ProductEntity.java
@@ -30,7 +30,7 @@ public class ProductEntity {
     @Column(name = "total_review")
     private Integer totalReview;
     @Column(name = "rating")
-    private Integer rating;
+    private float rating;
     @Column(name = "image")
     private String image;
     @Column(name = "attach_files")
